Fix excessive memory usage and CPU load in HLT TPC track merger while flagging shared clusters when clusters ID are not contiguous

+++ GPU/TPCCAGPUTracking/SliceTracker/AliHLTTPCCASettings.h
@@ -42,8 +42,8 @@
 typedef unsigned int calink;
 typedef unsigned int cahit;
 #else
-typedef unsigned short calink;
-typedef unsigned short cahit;
+typedef unsigned int calink;
+typedef unsigned int cahit;
 #endif
 
 #ifdef HLTCA_GPUCODE
